Implemented Role and User storage
Roles are stored in the database and cached locally when first seen. Remaining ToDo: Fill database automatically with all existing roles on the server (multiple servers?)

Users are only cached locally as of now. If no user is found in the cache the bot will search the server for it. ToDo: Cache users in database so we can get those at startup and provide us with a nice starter cache.